Removed url_name and use NS API provided names
The url_name column in the database, and the url_name attribute on
Nation and Region objects in the nationstates module have been removed.

They were a stop-gap measure, and now we query the NS API to ensure a
region actually exists, and to get the 'cannonical' name for the NS API.

We still need to normalize the name for comparisons, but this makes
things easier and removes some sketchiness. I've also added in a way to
get the region from the NS API.