Additional refactoring of code.
The code has been massively refactored, this is to make it easier to manage and update, as it was kind-of hard to read and unwieldy.

The enums.py file has been removed as we stopped using it with the new layout. Additionally, some exceptions have been removed because we stopped using them.

All of the NSVerify-specific functions and methods are now in the proper module, and I've taken the time to remove a good bit of the code that we had before, which makes it much easier to read.

Most of the modules have been reformatted as well, as necessary.